Fixed a bug in offline transactions
Fixed a bug in offline transactions that was making it so that if a
transaction does not go through (ex. amount is >r300) the results screen
was showing the results of the past transaction.  Made the
TransactionService.lastTransaction=errorMsg; instead of not updating
from the previous transaction (a potential security vulnerability)