The possibility of a later test completing before an earlier one gets done is now very unlikely
Although I believe the possibility of a later test completing before an
earlier one gets done is now very unlikely, I am not 100% done with the
modifications to the test.js files because I would like to harness the
powers of Protractor to simulate things like button clicks and input
auto completion.  So hold off on making the modifications to the php
code until I can get both sample tests working so that I know that
rerunning the code generation is not going to screw up things.